add middleware and refactoring
