+++ src/actions/endpoinds.ts
@@ -1,3 +1,3 @@
-export const BASE_URL = "https://now-qkewrwpxhm.now.sh/";
+export const BASE_URL = "https://bconf.xyz/";
 export const AUTH_URL = `${BASE_URL}auth/login`;
 export const SIGN_UP_URL = `${BASE_URL}auth/registry`;
