+++ src/constants/actions.ts
@@ -9,5 +9,6 @@ export const SIGN_UP_ERROR = "SIGN_UP_ERROR";
 
 // chat
 export const NEW_MESSAGE = "NEW_MESSAGE";
+export const OUTCOMING_MESSAGE = "OUTCOMING_MESSAGE";
 export const SEND_MESSAGE = "SEND_MESSAGE";
 export const RECEIVE_MESSAGE = "RECEIVE_MESSAGE";
